+++ lib/parseOptions.js
@@ -2,52 +2,12 @@ import { resolve } from 'path'
 
 import moment from 'moment'
 
+import { getHeadersConfig } from './utils/headers'
 import { proxyStringToObject, agentFromProxy } from 'contentful-batch-libs/dist/proxy'
 
 import { version } from '../package'
 
 import qs from 'querystring'
-
-/**
- * Turn header option into an object. Invalid header values
- * are ignored.
- *
- * @example
- * getHeadersConfig('Accept: Any')
- * // -> {Accept: 'Any'}
- *
- * @example
- * getHeadersConfig(['Accept: Any', 'X-Version: 1'])
- * // -> {Accept: 'Any', 'X-Version': '1'}
- *
- * @param value {string|string[]}
- */
-function getHeadersConfig (value) {
-  if (!value) {
-    return {}
-  }
-
-  const values = Array.isArray(value) ? value : [value]
-
-  return values.reduce((headers, value) => {
-    value = value.trim()
-
-    const separatorIndex = value.indexOf(':')
-
-    // Invalid header format
-    if (separatorIndex === -1) {
-      return headers
-    }
-
-    const headerKey = value.slice(0, separatorIndex).trim()
-    const headerValue = value.slice(separatorIndex + 1).trim()
-
-    return {
-      ...headers,
-      [headerKey]: headerValue
-    }
-  }, {})
-}
 
 export default function parseOptions (params) {
   const defaultOptions = {
